+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Send project root path with events ([#1739](https://github.com/getsentry/sentry-dotnet/pull/1739))
+
 ### Fixes
 
 - Fix error with `ConcurrentHashMap` on Android <= 9 ([#1761](https://github.com/getsentry/sentry-dotnet/pull/1761))